Makefile: allow user CFLAGS to override out warning flags
The cc-option filters out unknown flags which make the compiler error
out.
Note that Clang enables Wunused-command-line-argument by default and
merely emits a warning but incurs no error, when it encounters such a
flag.  GCC on the other hand exits unsuccessfully when it is passed a
flag of the form -Wunknown-warning, but no diagnostic is produced for
-Wno-unknown-warning unless other diagnostics are produced.